Buttons, Encapsulated Scripts, and Thread Changes
I have added code to improve the GUIButtons so that they can use EncapsulatedScript objects that I created. These objects come with code that can be executed by Buttons or any aspect of the program. To make the PauseScript possible, I had to modify all of the game's Thread-based code to work well with it. The PauseScript is a script that either pauses or resumes the game, depending on whether the object carries a true or false value, signifying pause and resume, respectively.