Fixes and Peripheral Spawner
I created a new Spawner that will spawn Attack Helicopters outside of the user's field of view. While creating it, I noticed a few issues, the most noteworthy of which is that the helicopters would crash into the ground whether there was a pilot or not. I also fixed a few method names and functions in a few classes, and added a tiny amount of documentation here and there.