Team-Colored Buildings
Buildings will have an outline for the team they are owned by, if they are owned by a particular team. Buildings can also be captured for a team if they can control the nearby area for long enough. I've also encapsulated some code and made small additions to accommodate it.